requirements.txt
asgiref==3.8.0
dj-rest-auth==5.0.2
Django==5.0.3
djangorestframework==3.15.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.10.0
